+++ app/components/Form/user/Profile.tsx
@@ -1,3 +1,4 @@
+import {useForm} from 'react-hook-form'
 import {useAuthState} from '@hooks/auth/useAuthState'
 
 import UserAvatar from '@components/User/Avatar'
@@ -13,19 +14,18 @@ import Textarea from '@components/form/Textarea/Textarea'
 import {SubmitButton} from '@components/form/Button/Submit'
 import CancelButton from '@components/form/Button/Cancel'
 
-import type {IProfileForm} from 'app/types/form'
+import type {IUserContext} from 'app/types/user'
 
 const ProfileForm = ({
-  handleSubmit,
-  onSubmit,
-  register,
-  errors,
-  reset,
-  control,
-  setValue,
   token,
-}: IProfileForm): JSX.Element => {
+  onSubmit,
+}: {
+  token: string
+  onSubmit: (data: IUserContext) => Promise<IUserContext>
+}): JSX.Element => {
   const {user} = useAuthState()
+  const {handleSubmit, register, errors, control, setValue, reset} =
+    useForm<IUserContext>()
   return (
     <form
       onSubmit={handleSubmit(onSubmit)}
