+++ app/components/User/Username.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react'
-import {useForm} from 'react-hook-form'
 import useEditUsername from '@hooks/user/useEditUsername'
 
 import Panel from '@components/route/Tablist/Panel'
@@ -7,21 +6,13 @@ import UsernameForm from '@components/Form/user/Username'
 import SnackBar from '@components/lib/SnackBar/SnackBar'
 
 import type {PanelPropsType} from 'app/types/components'
-import type {IEditUsername} from 'app/types/user'
 
 const EditUsername = ({token, isSelectedTab}: PanelPropsType): JSX.Element => {
-  const {handleSubmit, register, errors, reset} = useForm<IEditUsername>()
   const [isSuccess, successMessage, handleClose, onSubmit] =
     useEditUsername(token)
   return (
     <Panel index={1} isSelectedTab={isSelectedTab}>
-      <UsernameForm
-        handleSubmit={handleSubmit}
-        onSubmit={onSubmit}
-        register={register}
-        errors={errors}
-        reset={reset}
-      />
+      <UsernameForm onSubmit={onSubmit} />
       {isSuccess && (
         <SnackBar
           color='green'
